Refactor recording package
